backend/requirements.txt
Django==5.1.8
djangorestframework==3.15.2
psycopg[binary]==3.2.6
celery[redis]==5.4.0
redis==5.2.1
django-cors-headers==4.7.0
dj-database-url==2.3.0